refactor : Enum 타입 포함된 builder 수정
불필요한 @JsonCreator, @JsonProperty 삭제
