+++ src/main/java/com/yht/exerciseassist/domain/accuse/dto/AccuseReq.java
@@ -1,10 +1,6 @@
 package com.yht.exerciseassist.domain.accuse.dto;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.yht.exerciseassist.domain.accuse.AccuseType;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
@@ -13,14 +9,13 @@ import lombok.Data;
 public class AccuseReq {
 
     @NotNull(message = "신고 타입을 선택해주세요.")
-    @Enumerated(EnumType.STRING)
+
     private AccuseType accuseType;
 
     @NotEmpty(message = "신고 내용을 입력해주세요.")
     private String content;
 
-    @JsonCreator
-    public AccuseReq(@JsonProperty("accuseType") AccuseType accuseType, @JsonProperty("content") String content) {
+    public AccuseReq(AccuseType accuseType, String content) {
         this.accuseType = accuseType;
         this.content = content;
     }
