+++ src/main/java/com/yht/exerciseassist/domain/comment/service/CommentService.java
@@ -1,6 +1,5 @@
 package com.yht.exerciseassist.domain.comment.service;
 
-import com.yht.exerciseassist.ResponseResult;
 import com.yht.exerciseassist.domain.DateTime;
 import com.yht.exerciseassist.domain.comment.Comment;
 import com.yht.exerciseassist.domain.comment.dto.WriteCommentDto;
@@ -10,6 +9,7 @@ import com.yht.exerciseassist.domain.member.repository.MemberRepository;
 import com.yht.exerciseassist.domain.post.Post;
 import com.yht.exerciseassist.domain.post.repository.PostRepository;
 import com.yht.exerciseassist.exception.error.ErrorCode;
+import com.yht.exerciseassist.util.ResponseResult;
 import com.yht.exerciseassist.util.SecurityUtil;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
