+++ src/main/java/com/yht/exerciseassist/domain/diary/dto/ExerciseInfoDto.java
@@ -1,10 +1,6 @@
 package com.yht.exerciseassist.domain.diary.dto;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.yht.exerciseassist.domain.diary.BodyPart;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -23,17 +19,12 @@ public class ExerciseInfoDto {
 
     private int cardioTime; //유산소 운동 시간
 
-    @Enumerated(EnumType.STRING)
     private BodyPart bodyPart;
 
     private boolean finished; //완료 여부
 
-    @JsonCreator
     @Builder
-    public ExerciseInfoDto(@JsonProperty("exerciseName") String exerciseName, @JsonProperty("reps") int reps,
-                           @JsonProperty("exSetCount") int exSetCount, @JsonProperty("cardio") boolean cardio,
-                           @JsonProperty("cardioTime") int cardioTime, @JsonProperty("bodyPart") BodyPart bodyPart,
-                           @JsonProperty("finished") boolean finished) {
+    public ExerciseInfoDto(String exerciseName, int reps, int exSetCount, boolean cardio, int cardioTime, BodyPart bodyPart, boolean finished) {
         this.exerciseName = exerciseName;
         this.reps = reps;
         this.exSetCount = exSetCount;
