+++ src/main/java/com/yht/exerciseassist/domain/post/PostType.java
@@ -3,5 +3,4 @@ package com.yht.exerciseassist.domain.post;
 public enum PostType {
     Q_AND_A, KNOWLEDGE, SHOW_OFF, COMPETITION, FREE
     //       지식 공유     자랑       평가
-
 }
