+++ src/main/java/com/yht/exerciseassist/domain/post/dto/WritePostDto.java
@@ -1,11 +1,7 @@
 package com.yht.exerciseassist.domain.post.dto;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.yht.exerciseassist.domain.post.PostType;
 import com.yht.exerciseassist.domain.post.WorkOutCategory;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
@@ -21,17 +17,13 @@ public class WritePostDto {
     private String content;
 
     @NotNull(message = "게시글 카테고리를 선택해주세요.")
-    @Enumerated(EnumType.STRING)
     private PostType postType;
 
     @NotNull(message = "운동 종류를 선택해주세요.")
-    @Enumerated(EnumType.STRING)
     private WorkOutCategory workOutCategory;
 
-    @JsonCreator
     @Builder
-    public WritePostDto(@JsonProperty("title") String title, @JsonProperty("content") String content,
-                        @JsonProperty("postType") PostType postType, @JsonProperty("workOutCategory") WorkOutCategory workOutCategory) {
+    public WritePostDto(String title, String content, PostType postType, WorkOutCategory workOutCategory) {
         this.title = title;
         this.content = content;
         this.postType = postType;
