+++ src/main/java/com/yht/exerciseassist/jwt/filter/ExceptionHandlerFilter.java
@@ -1,13 +1,13 @@
 package com.yht.exerciseassist.jwt.filter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.yht.exerciseassist.ResponseResult;
 import com.yht.exerciseassist.exception.dto.ExceptionResponse;
 import com.yht.exerciseassist.exception.error.jwt.EmptyJWTTokenExcep;
 import com.yht.exerciseassist.exception.error.jwt.ExpiredJWTTokenExcep;
 import com.yht.exerciseassist.exception.error.jwt.InvalidJWTTokenExcep;
 import com.yht.exerciseassist.exception.error.jwt.UnsupportedJWTTokenExcep;
 import com.yht.exerciseassist.jwt.service.JWTService;
+import com.yht.exerciseassist.util.ResponseResult;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
