+++ src/main/java/com/yht/exerciseassist/util/ResponseResult.java
@@ -1,4 +1,4 @@
-package com.yht.exerciseassist;
+package com.yht.exerciseassist.util;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
