+++ src/test/java/com/yht/exerciseassist/domain/accuse/service/AccuseServiceTest.java
@@ -1,6 +1,5 @@
 package com.yht.exerciseassist.domain.accuse.service;
 
-import com.yht.exerciseassist.ResponseResult;
 import com.yht.exerciseassist.domain.accuse.Accuse;
 import com.yht.exerciseassist.domain.accuse.dto.AccuseReq;
 import com.yht.exerciseassist.domain.accuse.repository.AccuseRepository;
@@ -10,6 +9,7 @@ import com.yht.exerciseassist.domain.factory.PostFactory;
 import com.yht.exerciseassist.domain.member.Member;
 import com.yht.exerciseassist.domain.post.Post;
 import com.yht.exerciseassist.domain.post.repository.PostRepository;
+import com.yht.exerciseassist.util.ResponseResult;
 import com.yht.exerciseassist.util.SecurityUtil;
 import jakarta.persistence.EntityManager;
 import org.assertj.core.api.Assertions;
