+++ src/test/java/com/yht/exerciseassist/domain/member/controller/MemberControllerTest.java
@@ -2,13 +2,13 @@ package com.yht.exerciseassist.domain.member.controller;
 
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.yht.exerciseassist.ResponseResult;
 import com.yht.exerciseassist.domain.factory.MemberFactory;
 import com.yht.exerciseassist.domain.member.dto.SignInRequestDto;
 import com.yht.exerciseassist.domain.member.dto.SignUpRequestDto;
 import com.yht.exerciseassist.domain.member.service.MemberService;
 import com.yht.exerciseassist.exception.CommonExceptionHandler;
 import com.yht.exerciseassist.jwt.dto.TokenInfo;
+import com.yht.exerciseassist.util.ResponseResult;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
